refactor: group client helpers in common
Contributes to NEX-2231.

Groups client helpers in the `common` package to loose v1 client dependency in service packages.

- Fixes `IsUnknownRole` client helper: adds v1 client errors handling
- Removes `billing_group.go` — the file wasn't properly removed
